requirements.txt

selenium
PyGithub